requirements.txt
qdrant-edge-py
fastembed
fastapi
uvicorn
openai